Kept only ci_sanity tests for basic policy tests
